move RemoveUserFromAgency from connected-user to agency domain
